// blog-gui/src/app/open-api/model/post-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TagDTO } from './tag-dto';


export interface PostDTO { 
    postId?: string;
    title?: string;
    content?: string;
    userId?: string;
    fullName?: string;
    status?: string;
    tags?: TagDTO[];
}